% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pborepo-package.R
\docType{data}
\name{bsmmeta3}
\alias{bsmmeta3}
\title{BSM metadata 3}
\description{
BSM metadata 3
}
\details{
This is the preferred metadata source
}
\references{
\url{https://www.unavco.org/data/strain-seismic/bsm-data/metadata/metadata.html}
}
