% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unavco_methods.R
\name{consistent}
\alias{consistent}
\alias{consistent.default}
\alias{consistent.unavco}
\alias{consistent.zoo}
\title{Make the downloaded data consistent in time}
\usage{
consistent(dat, ...)

\method{consistent}{default}(dat, ...)

\method{consistent}{unavco}(dat, ...)

\method{consistent}{zoo}(dat, verbose = TRUE, ...)
}
\arguments{
\item{dat}{downloaded data to fortify}

\item{...}{additional parameters}
}
\description{
Make the downloaded data consistent in time
}
\examples{
consistent(1:10)  # does nothing but return

# ensures all possible data points are represented
library(zoo)
xc <- consistent(zoo(1:9, c(1,3:10))) # note the NA at index 2
na.approx(xc) # interpolate, for example
}
\seealso{
\code{\link{unavco_dataload}} and \code{\link{unavco-methods}}
}

