\name{describe}
\alias{describe}
\alias{describe.cal.pbo}
\title{Prints a description of the calibration table}
\usage{
  describe(tbl, ...)

  \method{describe}{cal.pbo} (tbl, ...)
}
\arguments{
  \item{tbl}{x}

  \item{...}{additional arguments}
}
\description{
  Prints a description of the calibration table
}

