% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calibrations.R
\name{describe_this}
\alias{describe_this}
\alias{describe_this.default}
\title{Prints a description of the calibration table}
\usage{
describe_this(tbl, ...)

\method{describe_this}{default}(tbl, ...)
}
\arguments{
\item{tbl}{the calibration table to describe}

\item{...}{additional arguments passed to \code{\link{cat}}}
}
\description{
Prints a description of the calibration table
}
\seealso{
Other Calibrations: \code{\link{all_calibrations}},
  \code{\link{all_calibrations.cal.hodg}},
  \code{\link{all_calibrations.cal.pbo}},
  \code{\link{all_calibrations.cal.pboiso}},
  \code{\link{any_calibration}},
  \code{\link{any_calibration.cal.hodg}},
  \code{\link{any_calibration.cal.pbo}},
  \code{\link{any_calibration.cal.pboiso}},
  \code{\link{any_calibration.default}},
  \code{\link{calibrate}}, \code{\link{calibration}},
  \code{\link{calibration_matrix}},
  \code{\link{calibration_matrix.cal.hodg}},
  \code{\link{calibration_matrix.cal.pbo}},
  \code{\link{calibration_matrix.cal.pboiso}},
  \code{\link{calibration_matrix.cal.roel}},
  \code{\link{fortify_calibration_matrix}};
  \code{\link{get_caltbl}},
  \code{\link{get_caltbl.default}}
}

