\name{get_caltbl}
\alias{get_caltbl}
\alias{get_caltbl.default}
\title{Get a specific calibration table}
\usage{
  get_caltbl(tblname, describe.tbl = TRUE, ...)

  \method{get_caltbl}{default} (tblname,
    describe.tbl = TRUE, ...)
}
\arguments{
  \item{tblname}{x}

  \item{...}{additional arguments}
}
\description{
  Get a specific calibration table
}

