% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{naOP}
\alias{fillNA}
\alias{interpNA}
\alias{locfNA}
\alias{naOP}
\alias{naOP.default}
\title{Automatically decide how to best fill in NA sections}
\usage{
naOP(d, fill = 0, verbose = TRUE, ...)

\method{naOP}{default}(d, fill = 0, verbose = TRUE, ...)

fillNA(d, fill = 0, ...)

locfNA(d, ...)

interpNA(d, ...)
}
\description{
Automatically decide how to best fill in NA sections
}
\seealso{
\code{\link{unavco-methods}}
}

