% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pbo_message}
\alias{pbo_message}
\title{Consistent messaging for downloading and other functions}
\usage{
pbo_message(..., lead.char = c("+", ">", "!"), n = 8)
}
\arguments{
\item{...}{the messages to give to \code{\link{message}}}

\item{lead.char}{character; the leading character}

\item{n}{numeric; the number of \code{lead.char} to replicate}
}
\description{
Consistent messaging for downloading and other functions
}
\seealso{
Other utilities: 
\code{\link{RMS}()},
\code{\link{list.txtgz}()},
\code{\link{norms}()}
}
\concept{utilities}
