% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pborepo-package.R
\docType{data}
\name{pborepo-constants}
\alias{pborepo-constants}
\alias{.pbo_constants}
\alias{constants}
\alias{pbo_constants}
\title{Constants used as defaults}
\format{
An object of class \code{list} of length 3.
}
\usage{
.pbo_constants

constants(do.str = TRUE)

pbo_constants()
}
\arguments{
\item{do.str}{logical; should the structure be printed?}
}
\description{
Constants used as defaults
}
\details{
The helper function \code{\link{constants}}
shows (the structure of, optionally)
and returns \code{.pbo_constants}. 
\code{\link{pbo_constants}} simply accesses \code{\link{constants}}
}
\examples{
str(pbo_constants())
}
\seealso{
\code{\link{pborepo}}
}
\keyword{datasets}
