\docType{package}
\name{pborepo-package}
\alias{pborepo}
\alias{pborepo-package}
\alias{strainmeter}
\title{Access to data-products and metadata from the Plate Boundary Observatory geodetic network}
\description{
  Some description
}
\details{
  See the following pages for details on
}
\section{Pore-pressure data}{
  \url{http://pore.unavco.org/pore}
}

\section{Pore-temperature data}{
  \url{http://pore.unavco.org/pore}
}
\author{
  Andrew J. Barbour <andy.barbour@gmail.com>
}

