% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pborepo_gitdata}
\alias{list_github_files}
\alias{pborepo_gitdata}
\title{Retrieve data from the online repository}
\usage{
pborepo_gitdata(file = NULL, save.local = TRUE, saf = FALSE,
  verbose = TRUE, ...)

list_github_files()
}
\arguments{
\item{file}{character;}

\item{save.local}{logical; should the file be saved locally}

\item{saf}{logical; the value of \code{stringsAsFactors}}

\item{verbose}{logical;}

\item{...}{additional parameters to \code{\link{read.https}}}
}
\description{
Retrieve data from the online repository
}

