% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/https.R
\name{read.https}
\alias{read.https}
\alias{get_https}
\title{Use \code{read.table} with a secure http url (https)}
\usage{
read.https(URL, local.file = NULL, delete = TRUE, verbose = TRUE,
  FUN = read.table, ...)

get_https(URL, ...)
}
\arguments{
\item{URL}{character; the location of the file}

\item{local.file}{character; if \code{NULL} a temporary file is used}

\item{delete}{logical; should \code{local.file} be delete upon exit?}

\item{verbose}{logical; should messages be shown?}

\item{FUN}{the function to use for reading \code{URL}; default is \code{\link{read.table}}}

\item{...}{additional parameters sent to \code{URL} or \code{\link{content}}}
}
\description{
Uses functions in the \pkg{httr} package to load
a file at a secure url (regular urls work too)
}
\details{
The main function to be used is \code{\link{read.https}}.
}
\examples{
\dontrun{
x <- "https://raw.githubusercontent.com/abarbour/pborepo/master/data/gen/lsm_coords"
get_https(x) # returns raw content
read.https(x, header=TRUE) # an actual data.frame instead
# (if 'x' is, say, a csv file, change FUN: read.https(x, FUN=read.csv)
}
}
\author{
A.J. Barbour
}
