% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trimmer}
\alias{trimmer}
\title{Trim a series by some percentage}
\usage{
trimmer(x, prc = 0.1)
}
\arguments{
\item{x}{series to trim}

\item{prc}{the percentage of data to trim from either side of \code{x}; this
cannot be less than \code{0} or greater than \code{0.5}.}
}
\description{
Trim a series by some percentage
}
\details{
The trimming is done with \code{\link{sort.int}}, which may be
problematic for timeseries.  The best option is to trim an index series
and use that to re-sample the timeseries.  See Examples.

Also sets an attribute \code{'trim'} with details of the trimming.
}
\examples{
# default is a 10\% trim
trimmer(1:10)  # 2 3 4 5 6 7 8 9

# but sort is used...
set.seed(1234)
x <- round(runif(10, max = 10)) # note order of values
trimmer(x) # trimmed, but sorted

# we can fix this by trimming an index:
x[trimmer(seq_along(x))] # trimmed, with order retained

}

