% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unavco-methods}
\alias{as.data.frame.unavco}
\alias{as.zoo.unavco}
\alias{data.frame.unavco}
\alias{unavco-methods}
\alias{window.unavco}
\alias{zoo.unavco}
\title{Methods for the 'unavco' class}
\usage{
as.zoo.unavco(x, frequency = NULL, ...)

zoo.unavco(x, frequency = NULL, ...)

\method{as.data.frame}{unavco}(x, ...)

data.frame.unavco(x, ...)

\method{window}{unavco}(x, ...)
}
\description{
Methods for the 'unavco' class
}
\seealso{
\code{\link{consistent}} and \code{\link{unavco_dataload}}
}

