\docType{data}
\name{bsmCalibrations}
\alias{bsmCalibrations}
\title{BSM calibration coefficients}
\format{A list with lists of dataframes with
calibration coefficients by station}
\description{
  BSM calibration coefficients
}
\examples{
load("bsmCalibrations.rda")
summary(bsmCalibrations)
#      Length Class  Mode
#grant 2      -none- list
#hodg  4      -none- list
#pbo   1      -none- list
#roel  2      -none- list
}
\keyword{datasets}

