% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pborepo-package.R
\docType{data}
\name{bsmCalibrations}
\alias{bsmCalibrations}
\title{BSM calibration coefficients}
\format{A list with lists of dataframes with
calibration coefficients by station}
\description{
BSM calibration coefficients
}
\examples{
data(bsmCalibrations)
summary(bsmCalibrations)
#      Length Class  Mode
#grant 2      -none- list
#hodg  4      -none- list
#pbo   1      -none- list
#roel  2      -none- list
}

