% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations.R
\name{calibration_matrix}
\alias{calibration_matrix}
\alias{calibrate}
\alias{calibration}
\alias{calibration_matrix.cal.pbo}
\alias{calibration_matrix.cal.pboiso}
\alias{calibration_matrix.cal.hodg}
\alias{calibration_matrix.cal.roel}
\alias{fortify_calibration_matrix}
\alias{all_calibrations}
\alias{all_calibrations.cal.hodg}
\alias{all_calibrations.cal.pbo}
\alias{all_calibrations.cal.pboiso}
\alias{any_calibration}
\alias{any_calibration.default}
\alias{any_calibration.cal.hodg}
\alias{any_calibration.cal.pbo}
\alias{any_calibration.cal.pboiso}
\title{Load or manipulate calibration matrices}
\usage{
calibration_matrix(tbl, sta4, typ = c("free", "cdr", "cd"), ...)

\method{calibration_matrix}{cal.pbo}(tbl, sta4, typ = c("free", "cdr", "cd"), ...)

\method{calibration_matrix}{cal.pboiso}(tbl, sta4, typ = c("free", "cdr", "cd"), ...)

\method{calibration_matrix}{cal.hodg}(tbl, sta4, typ = NULL, ...)

\method{calibration_matrix}{cal.roel}(tbl, sta4, typ = NULL, ...)

fortify_calibration_matrix(
  Sij,
  sta4 = NA,
  typ = NA,
  needs.pinv = FALSE,
  byrow = TRUE,
  ...
)

all_calibrations(tbl, sta4, ...)

\method{all_calibrations}{cal.hodg}(tbl, sta4, ...)

\method{all_calibrations}{cal.pbo}(tbl, sta4, ...)

\method{all_calibrations}{cal.pboiso}(tbl, sta4, ...)

any_calibration(tbl, sta4, preference, ...)

\method{any_calibration}{default}(tbl, sta4, preference = c("free", "cdr", "cd"), ...)

\method{any_calibration}{cal.hodg}(tbl, sta4, preference = c("free", "cdr", "cd"), ...)

\method{any_calibration}{cal.pbo}(tbl, sta4, preference = c("free", "cdr", "cd"), ...)

\method{any_calibration}{cal.pboiso}(tbl, sta4, preference = c("free", "cdr", "cd"), ...)
}
\arguments{
\item{tbl}{character; the name of the table to load; defaults to \code{"pbo"}
if missing.  The function will dispatch the appropriate method depending on the
class of the object}

\item{sta4}{character; the four-character station code, e.g. \code{'B084'}}

\item{typ}{character; an optional method identifier}

\item{...}{additional arguments}

\item{Sij}{matrix, or an object to be coerced into a matrix}

\item{needs.pinv}{logical; indicate whether the matrix should be pseudo-inverted}

\item{byrow}{logical; if \code{Sij} needs to be coerced into a matrix, should
it be filled `byrow'?}

\item{preference}{character; a vector of preferred calibration types (decreasing 
in preference with increasing index number); the first valid calibration matrix
is returned based on this preference}
}
\description{
Load or manipulate calibration matrices
}
\examples{
\dontrun{
#
# select a calibration table
#  data(bsmCalibrations)
#  pbotbl <- bsmCalibrations[['pbo']]
#  describe_this(pbotbl)
# or simply
pbotbl <- get_caltbl("pbo")
# 
# indiviual tables
calibration_matrix(pbotbl, "B084")
calibration_matrix(pbotbl, "B082") # Full of NAs (no calib.)
#
# All of the available calibrations (note the loss of attributes)
all_calibrations(pbotbl, "B084") # all three are available
all_calibrations(pbotbl, "B082") # but there's at least one
#
# First available calibration
any_calibration(pbotbl, "B084")
#
}
}
\seealso{
\code{\link{pinv}} to calculate the pseudoinverse

\code{\link{station_data}} to find information about the PBO stations (including
\code{sta4})

Other Calibrations: 
\code{\link{describe_this}()},
\code{\link{get_caltbl}()}
}
\author{
A. Barbour
}
\concept{Calibrations}
