% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data-retrieval}
\alias{data-retrieval}
\alias{dlCatcher}
\alias{pore.pressure}
\alias{pore.temperature}
\alias{unavco_dataload}
\alias{unavco_dataload.default}
\alias{unavco_downloader}
\alias{unavco_downloader.default}
\alias{unavco_path}
\alias{unavco_temploader}
\alias{unavco_temploader.default}
\title{Retrieve and load data from the UNAVCO archive}
\usage{
pore.pressure(sta4, year, jday, dt.window = NULL, ...)

pore.temperature(sta4, year, jday, dt.window = NULL, ...)

unavco_dataload(sta, year, jday, dattype = c("pp", "pt"),
  default.sec.samp = 1, localfile = NULL, quiet = FALSE, impute = TRUE)

\method{unavco_dataload}{default}(sta, year, jday, dattype = c("pp", "pt"),
  default.sec.samp = 1, localfile = NULL, quiet = FALSE, impute = TRUE)

unavco_temploader(file, dattype = c("pp", "pt"), ...)

\method{unavco_temploader}{default}(file, dattype = c("pp", "pt"), ...)

unavco_downloader(url.toget, temp, PLUGIN, quiet = FALSE,
  remove.temp = FALSE, ...)

\method{unavco_downloader}{default}(url.toget, temp, PLUGIN, quiet = FALSE,
  remove.temp = FALSE, ...)

dlCatcher(FUN)

unavco_path(sta4, year, jday, type = c("pp", "pt"), url.check = FALSE)
}
\arguments{
\item{sta}{character;}

\item{dattype}{character;}

\item{default.sec.samp}{numeric;}

\item{localfile}{character;}

\item{impute}{logical;}

\item{sta4}{character; 4-character station name (e.g. B084).}

\item{year}{numeric; the year of data}

\item{jday}{numeric; Julian date (the day of \code{year})}

\item{dt.window}{xx; extract data from this window; should be \code{POSIXt} compatible}

\item{type}{character; the type of data to generate download data for}

\item{url.check}{logical; should the urls be tested for existence?}

\item{url.toget}{character; the URL to download}

\item{temp}{character; file for temporary writing}

\item{quiet}{logical; should the functions not be verbose?}

\item{file}{character; the file to be loaded}

\item{PLUGIN}{function (not used)}

\item{FUN}{a function to try, and catch exceptions from}

\item{remove.temp}{logical; should temporary files be deleted?}

\item{...}{additional parameters}
}
\description{
The user should be most interested in \code{\link{pore.pressure}} and
\code{\link{pore.temperature}} to get pore-fluid pressure and temperature
data.
}
\details{
\code{\link{pore.pressure}} and \code{\link{pore.temperature}} both use
\code{\link{unavco_dataload}} and \code{\link{unavco_dataload}}  to
download, reads-in, and process data from the UNAVCO archive.

These functions will return all data for the specified year and (Julian) day(s).

\code{\link{unavco_temploader}} loads and process data downloaded from the UNAVCO archive

\code{\link{unavco_downloader}} is the downloader for the UNAVCO data archive
\code{\link{download.file}} is used with \code{cacheOK==FALSE}, and \code{mode=="w"}.

\code{\link{dlCatcher}} handles exceptions.

\code{\link{unavco_path}} generates the appropriate path for
UNAVCO data retrieval; it returns the appropriate url to download a .txt.gz data file
from the UNAVCO data archive.
If \code{url.check==TRUE} the function \code{url.exists} is used
to test the status of each link; this can add a substantial amount
of processing time for large numbers of urls.
}
\examples{
\dontrun{
# Download data, also makes it consistent
xpp<-pore.pressure("B084",2010,94)
xpt<-pore.temperature("B084",2010,94)

layout(matrix(1:2))
plot(xpp, type='l')
plot(xpt, type='l')
layout(matrix(1))
}
unavco_path("B084", 2010, 1)
unavco_path("B084", 2010, 1:10) # use vectors for multiple urls
unavco_path("B084",2010, 1:10, "pt")
# Also try:
# unavco_path("B084",2010, 1:10, "pt", TRUE)
}
\seealso{
\code{\link{read.https}} and XXX to load data files from
the source (github)

\code{\link{consistent}}

\code{\link{unavco-methods}}
}

