\name{describe}
\alias{describe}
\alias{describe.cal.pbo}
\title{Prints a description of the calibration table}
\usage{
describe(tbl, ...)

\method{describe}{cal.pbo}(tbl, ...)
}
\arguments{
  \item{tbl}{the calibration table to describe}

  \item{...}{additional arguments passed to
  \code{\link{cat}}}
}
\description{
Prints a description of the calibration table
}
\seealso{
Other Calibrations: \code{\link{all_calibrations}},
\code{\link{all_calibrations.cal.pbo}},
\code{\link{all_calibrations.cal.pbo}},
\code{\link{any_calibration}},
\code{\link{any_calibration.cal.pbo}},
\code{\link{any_calibration.cal.pbo}},
\code{\link{calibration_matrix}},
\code{\link{calibration_matrix.cal.hodg}},
\code{\link{calibration_matrix.cal.hodg}},
\code{\link{calibration_matrix.cal.pbo}},
\code{\link{calibration_matrix.cal.pbo}},
\code{\link{calibration_matrix.cal.roel}},
\code{\link{calibration_matrix.cal.roel}},
\code{\link{fortify_calibration_matrix}};
\code{\link{get_caltbl}}, \code{\link{get_caltbl.default}},
\code{\link{get_caltbl.default}}
}

