% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations.R
\name{describe_this}
\alias{describe_this}
\alias{describe_this.default}
\title{Prints a description of the calibration table}
\usage{
describe_this(tbl, ...)

\method{describe_this}{default}(tbl, ...)
}
\arguments{
\item{tbl}{the calibration table to describe}

\item{...}{additional arguments passed to \code{\link{cat}}}
}
\description{
Prints a description of the calibration table
}
\seealso{
Other Calibrations: 
\code{\link{calibration_matrix}()},
\code{\link{get_caltbl}()}
}
\concept{Calibrations}
