% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrations.R
\name{get_caltbl}
\alias{get_caltbl}
\alias{get_caltbl.default}
\title{Get a specific calibration table}
\usage{
get_caltbl(tblname, describe.tbl = TRUE, ...)

\method{get_caltbl}{default}(tblname, describe.tbl = TRUE, ...)
}
\arguments{
\item{tblname}{character; the name of the calibration table
to import. See \code{\link{bsmCalibrations}}.  Defaults to
\code{"pbo"} if missing.}

\item{describe.tbl}{logical; should \code{\link{describe_this}}
be used to print information about the calibration table?}

\item{...}{additional arguments}
}
\description{
Get a specific calibration table
}
\seealso{
\code{\link{bsmCalibrations}}

Other Calibrations: \code{\link{calibration_matrix}},
  \code{\link{describe_this}}
}
\concept{Calibrations}
