% Generated by roxygen2 (4.0.2): do not edit by hand
\name{norms}
\alias{norms}
\title{Calculate \emph{all} norms of a matrix}
\usage{
norms(M, types = NULL)
}
\arguments{
\item{M}{matrix}

\item{types}{character; optionally, specify the norms you wish to calculate
e.g., O, I, F, M, 2}
}
\description{
Calculate \emph{all} norms of a matrix
}
\details{
\code{\link{na.omit}} is used on \code{M} prior to calculation
}
\examples{
M <- matrix(1:12,4)
norms(M)
#
# or manually specify
norms(M, c("O","M"))
}
\seealso{
\code{\link{norm}}

Other utilities: \code{\link{.location.names}},
  \code{\link{Anza_stations}}, \code{\link{location}},
  \code{\link{sta16_from_sta4}},
  \code{\link{station_data}}; \code{\link{RMS}},
  \code{\link{colRMS}}, \code{\link{rms}},
  \code{\link{rowRMS}}; \code{\link{list.txtgz}};
  \code{\link{pbo_message}}
}

