% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pbo_message}
\alias{pbo_message}
\title{Consistent messaging for downloading and other functions}
\usage{
pbo_message(..., lead.char = c("+", ">", "!"), n = 8)
}
\arguments{
\item{...}{the messages to give to \code{\link{message}}}

\item{lead.char}{character; the leading character}

\item{n}{numeric; the number of \code{lead.char} to replicate}
}
\description{
Consistent messaging for downloading and other functions
}
\seealso{
Other utilities: \code{\link{.location.names}},
  \code{\link{location}}, \code{\link{sta16_from_sta4}},
  \code{\link{station_data}}; \code{\link{RMS}},
  \code{\link{colRMS}}, \code{\link{rms}},
  \code{\link{rowRMS}}; \code{\link{list.txtgz}};
  \code{\link{norms}}
}

