\docType{data}
\name{pborepo-constants}
\alias{pbo_constants}
\alias{pborepo-constants}
\title{Constants used as defaults}
\format{List of 3
 $ unavco     :List of 5
 $ conversions:List of 4
 $ bsm        :List of 5}
\usage{
  .constants

  pbo_constants(do.str = TRUE)
}
\arguments{
  \item{do.str}{logical; should the structure be printed?}
}
\description{
  Constants used as defaults
}
\details{
  The helper function \code{\link{pbo_constants}} shows
  (the structure of, optionally) and returns
  \code{.constants}.
}
\seealso{
  \code{\link{pborepo}}

  Other overviews: \code{\link{pborepo-package}}
}
\keyword{datasets}

