\docType{package}
\name{pborepo-package}
\alias{pborepo}
\alias{pborepo-package}
\title{Access to data-products and metadata from the Plate Boundary Observatory geodetic network}
\description{
Some description
}
\details{
See the following pages for details on
}
\section{Pore-pressure data}{
  \url{http://pore.unavco.org/pore}
}

\section{Pore-temperature data}{
  \url{http://pore.unavco.org/pore}
}
\author{
Andrew J. Barbour <andy.barbour@gmail.com>
}
\seealso{
See \code{\link{data-retrieval}} for functions related to
data retrieval

See \code{\link{station_data}} for example, for utility
functions

See \code{\link{calibration_matrix}} for example, for
functions associated with strain calibration coefficients.
}

