% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/https.R
\name{pborepo_gitdata}
\alias{gitdata}
\alias{list_github_files}
\alias{pborepo_gitdata}
\title{Retrieve data from the \code{pborepo} online repository (github)}
\usage{
pborepo_gitdata(file = NULL, save.local = FALSE, saf = FALSE,
  verbose = TRUE, ...)

list_github_files()
}
\arguments{
\item{file}{character; the name of the file to retrieve; if this is \code{NULL} only the
results of \code{\link{list_github_files}} are returned}

\item{save.local}{logical; should the file be saved locally}

\item{saf}{logical; the value of \code{stringsAsFactors}}

\item{verbose}{logical;}

\item{...}{additional parameters to \code{\link{read.https}}}
}
\description{
Retrieve data from the \code{pborepo} online repository (github)
}
\examples{
\dontrun{
#
list_github_files() # returns the filenames, with other info
#
print(gitfiles <- pborepo_gitdata()) # returns only the filenames
print(lsm <- pborepo_gitdata("lsm_coords")) # returns a data.frame
pborepo_gitdata("bogus_filename") # an error, as expected
#
# Get them all with plyr/dplyr
library(plyr)
library(dplyr, warn.conflicts=FALSE)
alldata <- llply(gitfiles, function(fi) tbl_df(pborepo_gitdata(fi, verbose=FALSE)), .progress = 'text')
names(alldata) <- gitfiles
summary(alldata)
print(str(alldata, vec.len=2, nchar.max=10))
lsm2 <- alldata[['lsm_coords']]
all.equal(tbl_df(lsm),lsm2)
}
}
\seealso{
\code{\link{read.https}}, which is used to retrieve the data
}

