% Generated by roxygen2 (4.0.2): do not edit by hand
\name{station_data}
\alias{.location.names}
\alias{Anza_stations}
\alias{location}
\alias{sta16_from_sta4}
\alias{station_data}
\title{Station information from PBO metadata}
\usage{
station_data(name = NULL, meta = "bsm", use.regexp = FALSE,
  verbose = TRUE)

Anza_stations(...)

sta16_from_sta4(sta4 = NULL, meta = "bsm", use.regexp = FALSE)

location(sta4 = NULL, ...)

.location.names(x)
}
\arguments{
\item{name}{character; the search string; use \code{use.regexp=TRUE} to use it
as a regular expression}

\item{meta}{character; the metadata source; currently the only
option is \code{'bsm'}, which accesses
PBO-BSM station information (see \code{data('bsmmeta')})}

\item{use.regexp}{logical; should \code{name} be considered to be a regular expression
(c.f. \code{\link{regexp}})?}

\item{verbose}{logical; should messages be given?}

\item{sta4}{character; an optional 4-character station code (e.g., \code{'B084'})}
}
\description{
PBO operates a large number of stations, and \code{\link{station_data}} is
a mechanism to search through a subset of their metadata;
\code{\link{sta16_from_sta4}} is a simple way to convert from a station's
4-character station code (e.g., B084) to its
16-character representation (e.g., pinyon084bcs2006).
}
\details{
This function searches internal metadata, which can be accessed
via \code{data(bsmmeta)}, for example.
}
\examples{
\dontrun{
station_data()
station_data("B084")
station_data("bcs", use.regexp=TRUE)
station_data(c("B082","B084"), use.regexp=TRUE) # will use only the first
station_data("B08[24]", use.regexp=TRUE) # gets what you want
#
# Get all of the 16-character codes
sta16_from_sta4()
# or just a few
sta16_from_sta4(c("B082","B084"))

# Special wrapper to get Anza cluster stations
# (note the extra bit of information returned: has.pp)
Anza_stations()
}
}
\author{
A. Barbour
}
\seealso{
Other utilities: \code{\link{RMS}}, \code{\link{colRMS}},
  \code{\link{rms}}, \code{\link{rowRMS}};
  \code{\link{list.txtgz}}; \code{\link{norms}};
  \code{\link{pbo_message}}
}

